+++ fair/cli.py
@@ -31,20 +31,20 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 import click
 import typing
 
-from dante.services import download_data
-from dante.dante import DANTE
+from fair.services import download_data
+from fair.fair import FAIR
 
 __author__ = "Scottish COVID Response Consortium"
 __credits__ = ["Nathan Cummings (UKAEA)", "Kristian Zarebski (UKAEA)"]
 __license__ = "BSD-2-Clause"
 __status__ = "Development"
-__copyright__ = "Copyright 2021, DANTE"
+__copyright__ = "Copyright 2021, FAIR"
 
 
 @click.group()
 @click.version_option()
 def cli():
-    """Welcome to DANTE, the FAIR data pipeline command-line interface."""
+    """Welcome to FAIR, the FAIR data pipeline command-line interface."""
     # if registry_installed() and registry_running():
     #     click.echo("Local registry installed and running")
     # else:
@@ -60,46 +60,46 @@ def cli():
 @cli.command()
 def status() -> None:
     """Get the status of files under staging"""
-    with DANTE() as s:
+    with FAIR() as s:
         s.status()
 
 
 @cli.command()
 def init() -> None:
     """Initialise repository in current location"""
-    with DANTE() as dante:
-        dante.initialise()
+    with FAIR() as fair:
+        fair.initialise()
 
 
 @cli.command()
 def purge() -> None:
     """resets the repository deleting all local caches"""
     _purge = click.prompt(
-        "Are you sure you want to reset dante tracking, "
+        "Are you sure you want to reset fair tracking, "
         "this is not reversible [Y/N]? ",
         type=click.BOOL,
     )
     if _purge:
-        with DANTE() as dante:
-            dante.purge()
+        with FAIR() as fair:
+            fair.purge()
 
 
 @cli.command()
 @click.argument("file_paths", type=click.Path(exists=True), nargs=-1)
 def reset(file_paths: typing.List[str]) -> None:
     """Removes files/runs from staging"""
-    with DANTE() as dante:
+    with FAIR() as fair:
         for file_name in file_paths:
-            dante.change_staging_state(file_name, False)
+            fair.change_staging_state(file_name, False)
 
 
 @cli.command()
 @click.argument("file_paths", type=click.Path(exists=True), nargs=-1)
 def add(file_paths: typing.List[str]) -> None:
     """Add a file to staging"""
-    with DANTE() as dante:
+    with FAIR() as fair:
         for file_name in file_paths:
-            dante.change_staging_state(file_name, True)
+            fair.change_staging_state(file_name, True)
 
 
 @cli.command()
@@ -111,9 +111,9 @@ def add(file_paths: typing.List[str]) -> None:
 )
 def rm(file_paths: typing.List[str], cached: bool = False) -> None:
     """removes files from system or just tracking"""
-    with DANTE() as dante:
+    with FAIR() as fair:
         for file_name in file_paths:
-            dante.remove_file(file_name, cached)
+            fair.remove_file(file_name, cached)
 
 
 @cli.command()
@@ -142,16 +142,16 @@ def pull(config: str):
 def run(ctx):
     """Initialises a run with the option to specify a bash command"""
     if not ctx.invoked_subcommand:
-        with DANTE() as dante:
-            dante.run_bash_command()
+        with FAIR() as fair:
+            fair.run_bash_command()
 
 
 @run.command()
 @click.argument("bash_command")
 def bash(bash_command: str):
     """Run a BASH command and set this to be the default run command"""
-    with DANTE() as dante:
-        dante.run_bash_command(bash_command)
+    with FAIR() as fair:
+        fair.run_bash_command(bash_command)
 
 
 @cli.group(invoke_without_command=True)
@@ -160,8 +160,8 @@ def bash(bash_command: str):
 def remote(ctx, verbose: bool = False):
     """typing.List remotes if no additional command is provided"""
     if not ctx.invoked_subcommand:
-        with DANTE() as dante:
-            dante.list_remotes(verbose)
+        with FAIR() as fair:
+            fair.list_remotes(verbose)
 
 
 @remote.command()
@@ -180,8 +180,8 @@ def add(options: typing.List[str]) -> None:
     _url = options[1] if len(options) > 1 else options[0]
     _label = options[0] if len(options) > 1 else "origin"
 
-    with DANTE() as dante:
-        dante.add_remote(_url, _label)
+    with FAIR() as fair:
+        fair.add_remote(_url, _label)
 
 
 @remote.command()
@@ -194,8 +194,8 @@ def remove(label: str) -> None:
     label : str
         label of remote to remove
     """
-    with DANTE() as dante:
-        dante.remove_remove(label)
+    with FAIR() as fair:
+        fair.remove_remove(label)
 
 
 @remote.command()
@@ -210,23 +210,23 @@ def modify(options: typing.List[str]) -> None:
     """
     _label = options[0] if len(options) > 1 else "origin"
     _url = options[1] if len(options) > 1 else options[0]
-    with DANTE() as dante:
-        dante.modify_remote(_label, _url)
+    with FAIR() as fair:
+        fair.modify_remote(_label, _url)
 
 
 @cli.command()
 def log() -> None:
     """Show a full run history"""
-    with DANTE() as dante:
-        dante.show_history()
+    with FAIR() as fair:
+        fair.show_history()
 
 
 @cli.command()
 @click.argument("run_id")
 def view(run_id: str) -> None:
     """View log for a given run"""
-    with DANTE() as dante:
-        dante.show_run_log(run_id)
+    with FAIR() as fair:
+        fair.show_run_log(run_id)
 
 
 @cli.command()
@@ -263,12 +263,12 @@ def config_user(user_name: str) -> None:
     (API token, associated namespace, local data
     store, login node, and so on).
     """
-    with DANTE() as dante:
-        dante.set_user(user_name)
+    with FAIR() as fair:
+        fair.set_user(user_name)
 
 
 @config.command(name="user.email")
 @click.argument("user_email")
 def config_email(user_email: str) -> None:
-    with DANTE() as dante:
-        dante.set_email(user_email)
+    with FAIR() as fair:
+        fair.set_email(user_email)